Suppression d'un article par son auteur

+++ src/control/UserControl/delete_article.php
@@ -10,8 +10,21 @@ if (empty($_SESSION["userID"]) || $_SESSION["userRole"] != "admin") {
     exit;
 }
 
-$id = htmlspecialchars($_GET["id"], ENT_QUOTES);
-$state = $bdd->prepare("DELETE FROM articles WHERE id = ?");
-$state->execute(array($id));
+if (empty($_SESSION["userID"])){
+    $articleID = htmlspecialchars($_GET["articleID"], ENT_QUOTES);
+    $firstAuthorId = $bdd->prepare("SELECT firstAuthor FROM articles WHERE id = ?");
+    $firstAuthorId->execute(array($articleID));  
+    $firstAuthorId->fetch();
+}
 
-header("Location: ../../../article_list.php");
+if(empty($_SESSION["userRole"] == "admin")  ||  (empty($_SESSION["userID"])== $firstAuthorId)){
+    $id = htmlspecialchars($_GET["id"], ENT_QUOTES);
+    $state = $bdd->prepare("DELETE FROM articles WHERE id = ?");
+    $state->execute(array($id));    
+}
+
+if(empty($_SESSION["userRole"] != "admin")){
+    header("Location: ../../../article_list.php");
+}else{
+    header("Location: ../../../home.php");  
+}
